backend/pyproject.toml
[project]
name = "logit-mail-backend"
version = "0.1.0"
description = "Backend for AI Email Assistant"
authors = [
    {name = "Manvith M", email = "manvithm@example.com"},
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
    "alembic>=1.11.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.9",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "chromadb>=0.4.22",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.1.0"
]
requires-python = ">=3.10"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
